+++ weex_core/Source/core/bridge/bridge.h
@@ -81,8 +81,8 @@ namespace WeexCore {
                                bool willLayout= true) = 0;
 
     virtual int callLayout(const char* pageId, const char* ref,
-                           int top, int bottom, int left, int right,
-                           int height, int width, int index) = 0;
+                           float top, float bottom, float left, float right,
+                           float height, float width, int index) = 0;
 
     virtual int callUpdateStyle(const char* pageId, const char* ref,
                                 std::vector<std::pair<std::string, std::string>> *style,
@@ -118,7 +118,7 @@ namespace WeexCore {
 
     void setDefaultHeightAndWidthIntoRootDom(const char *instanceId, const float defaultWidth, const float defaultHeight,
                                              const bool isWidthWrapContent, const bool isHeightWrapContent);
-    void setRenderContainerWrapContent(const char* instanceId,bool wrap);
+    void setRenderContainerWrapContent(const char* instanceId, bool wrap);
 
     /**Notice that this method is invoked from JS thread.**/
     void forceLayout(const char *instanceId);
